feat: add Instapaper integration
Save-for-later articles from Instapaper, synced via OAuth API.
Features: background sync, article download as Markdown,
Lyra-themed list UI sorted by date, and 60s WiFi cooldown on exit.
Settings UI for credentials with XOR-obfuscated consumer keys.

+++ src/activities/reader/MdReaderActivity.cpp
@@ -174,7 +174,7 @@ void MdReaderActivity::initializeReader() {
   orientedMarginBottom += cachedScreenMargin;
 
   orientedMarginBottom +=
-      std::max(cachedScreenMargin, static_cast<uint8_t>(UITheme::getStatusBarHeight()));
+      std::max(cachedScreenMargin, static_cast<int>(UITheme::getStatusBarHeight()));
 
   const uint16_t viewportWidth = renderer.getScreenWidth() - orientedMarginLeft - orientedMarginRight;
   const uint16_t viewportHeight = renderer.getScreenHeight() - orientedMarginTop - orientedMarginBottom;
@@ -369,7 +369,7 @@ void MdReaderActivity::renderScreen() {
   orientedMarginBottom += cachedScreenMargin;
 
   orientedMarginBottom +=
-      std::max(cachedScreenMargin, static_cast<uint8_t>(UITheme::getStatusBarHeight()));
+      std::max(cachedScreenMargin, static_cast<int>(UITheme::getStatusBarHeight()));
 
   auto page = loadPageFromCache(currentPage);
   if (!page) {
